Added files via upload
Added compatibility for run from windows iPython.
